Use emulated text tracks
As of video.js v5.14, video.js no longer supports native
subtitles on anything besides Safari. For Chrome, this means
there is *no* subtitle support. This will wait for dash.js to
load all the text tracks and then add them to video.js if we're
using a version that does not attempt to use native subtitles.